significantly reduce the API calls
Instead of using a search for every single field, retrieve all
of the existing student id records and all of the existing
student records in the app, and index them by student
name, so we can do a simple associative array lookup.
MUCH faster and no internet is needed at all for those.  Hopefully this
will fix the issues.